Moved the cookies page to common
