--- service-front/app/src/Actor/templates/actor/cookies.html.twig
@@ -1,66 +1,0 @@
-{% extends '@actor/layout.html.twig' %}
-
-{% block html_title %}Cookies{% endblock %}
-
-{% set usageCookies = form.get('usageCookies') %}
-
-{% block content %}
-    <div class="govuk-width-container">
-
-        <main class="govuk-main-wrapper" id="main-content" role="main">
-
-                    <h1 class="govuk-heading-xl">Cookies</h1>
-                    <p class="govuk-body">Cookies are files saved on your phone, tablet or computer when you visit a website.</p>
-                    <p class="govuk-body">We use cookies to store information about how you use the Make a lasting power of attorney service, such as the
-                        pages you visit.</p>
-
-                    {{ govuk_form_open(form) }}
-                    {{ govuk_form_element(form.get('__csrf')) }}
-
-                    <h2 class="govuk-heading-xl">Cookie settings</h2>
-                    <p class="govuk-body">We use two types of cookie. One type is strictly necessary and the other type of cookie is optional.</p>
-
-                    <h3 class="govuk-heading-m">Cookies that measure website use</h3>
-                    <p class="govuk-body">
-                        OPG and Government Digital Services use Google Analytics to measure how you use the service so we can improve it
-                        based on user needs. Google Analytics sets cookies that store anonymised information about:
-                    </p>
-
-                    <ul class="govuk-list govuk-list--bullet">
-                        <li>what pages you visit and how long you spend on each</li>
-                        <li>what you click on while you're visiting the site</li>
-                    </ul>
-
-                    <p class="govuk-body">
-                        We do not allow Google to use or share the data about how you use this site.
-                    </p>
-
-                    <div class="govuk-form-group">
-                        <fieldset class="govuk-fieldset" aria-describedby="changed-name-hint">
-                            <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
-                                <span class="govuk-visually-hidden question-text">Allow use of non-essential Cookies?</span>
-                            </legend>
-
-                            {{ govuk_form_element(form.get('usageCookies'), {'value_options':
-                                [
-                                    {
-                                        'label': 'Use cookies that measure my website use',
-                                        'value': 'yes'
-                                    },
-                                    {
-                                        'label': 'Do not use cookies that measure my website use',
-                                        'value': 'no'
-                                    }
-                                ]}
-                            )}}
-                        </fieldset>
-                    </div>
-                    <h3 class="govuk-heading-m">Strictly necessary cookies</h3>
-                    <p class="govuk-body">These essential cookies allow you to log in to the service. They always need to be on.</p>
-
-                    <button data-prevent-double-click="true" type="submit" class="govuk-button">Save changes</button>
-            {{ govuk_form_close() }}
-        </main>
-
-    </div>
-{% endblock %}
