+++ service-front/app/src/Common/templates/layout/default.html.twig
@@ -25,27 +25,27 @@
 <body class="govuk-template__body ">
 <a href="#main-content" class="govuk-skip-link">Skip to main content</a>
 
-<form>
-<div class="cookie-banner govuk-width-container" role="region" aria-label="cookie banner">
-    <div class="govuk-grid-row">
-        <div class=" govuk-grid-column-two-thirds">
-            <div class="cookie-banner__message">
-                <h2 class="govuk-heading-m">Tell us whether you accept cookies</h2>
-                <p class="govuk-body">GOV.UK uses cookies which are essential for the site to work. We also use non-essential cookies to help us
-                    improve government digital services. Any data collected is anonymised.</p>
-            </div>
-            <div class="cookie-banner__buttons">
-                <div class="cookie-banner__button cookie-banner__button-accept govuk-grid-column-full govuk-grid-column-one-half-from-desktop govuk-!-padding-left-0">
-                    <button class="govuk-button button--inline" name="accept-all-cookies" type="submit" role="button">Accept all cookies</button>
-                </div>
-                <div class="cookie-banner__button govuk-grid-column-full govuk-grid-column-one-half-from-desktop govuk-!-padding-left-0">
-                    <a href="/cookies" class="govuk-button govuk-button--secondary button--inline" name="set-cookie-preferences" role="button">Set cookie preferences</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-</form>
+{#<form>#}
+{#<div class="cookie-banner govuk-width-container" role="region" aria-label="cookie banner">#}
+{#    <div class="govuk-grid-row">#}
+{#        <div class=" govuk-grid-column-two-thirds">#}
+{#            <div class="cookie-banner__message">#}
+{#                <h2 class="govuk-heading-m">Tell us whether you accept cookies</h2>#}
+{#                <p class="govuk-body">GOV.UK uses cookies which are essential for the site to work. We also use non-essential cookies to help us#}
+{#                    improve government digital services. Any data collected is anonymised.</p>#}
+{#            </div>#}
+{#            <div class="cookie-banner__buttons">#}
+{#                <div class="cookie-banner__button cookie-banner__button-accept govuk-grid-column-full govuk-grid-column-one-half-from-desktop govuk-!-padding-left-0">#}
+{#                    <button class="govuk-button button--inline" name="accept-all-cookies" type="submit" role="button">Accept all cookies</button>#}
+{#                </div>#}
+{#                <div class="cookie-banner__button govuk-grid-column-full govuk-grid-column-one-half-from-desktop govuk-!-padding-left-0">#}
+{#                    <a href="/cookies" class="govuk-button govuk-button--secondary button--inline" name="set-cookie-preferences" role="button">Set cookie preferences</a>#}
+{#                </div>#}
+{#            </div>#}
+{#        </div>#}
+{#    </div>#}
+{#</div>#}
+{#</form>#}
 
 <header class="govuk-header " role="banner" data-module="header">
     <div class="govuk-header__container govuk-width-container">
